feat: 특정 시험 정보 조회 기능 구현

+++ src/main/java/com/example/bcsd_studium/domain/repository/ExamRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ExamRepository extends JpaRepository<Exam, Integer> {
+public interface ExamRepository extends JpaRepository<Exam, Long> {
     List<Exam> findAllByOrderByIdDesc();
 }
